requirements.txt
deta==1.2.0
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.5.0
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
scipy==1.11.1
streamlit==1.25.0
streamlit_authenticator==0.1.5
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
